# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API built with Golang using Clean Code architecture. \\nGithub: [https://github.com/HoangBD64/go-ecom].",
        "title": "E-commerce Application Backend API",
        "contact": {
            "name": "For API Support",
            "email": "nikhilnarayanan623@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "basePath": "/api",
    "paths": {
...
        "/account/wallet": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "API for user to get user wallet",
                "tags": [
                    "User Profile"
                ],
                "summary": "Get user wallet  (User)",
                "operationId": "GetUserWallet",
                "responses": {
                    "200": {
                        "description": "Successfully retrieve user wallet",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve user wallet",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/account/wallet/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "API for user to get user wallet transaction",
                "tags": [
                    "User Profile"
                ],
                "summary": "Get user wallet  (User)",
                "operationId": "GetUserWalletTransactions",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user wallet transactions",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "204": {
                        "description": "No wallet transaction for user",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve user wallet transactions",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Add prefix of Bearer before  token Ex: \"Bearer token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}